TG-28: corrected directory for projects tests
